+++ lib/datasource/cargo/index.js
@@ -1,4 +1,4 @@
-const got = require('../../util/got/datasource');
+const got = require('../../util/got');
 
 module.exports = {
   getPkgReleases,
