+++ lib/datasource/docker/index.js
@@ -4,7 +4,7 @@ const is = require('@sindresorhus/is');
 const parseLinkHeader = require('parse-link-header');
 const wwwAuthenticate = require('www-authenticate');
 
-const got = require('../../util/got/datasource');
+const got = require('../../util/got');
 const hostRules = require('../../util/host-rules');
 
 module.exports = {
