+++ lib/datasource/go/index.js
@@ -1,4 +1,4 @@
-const got = require('../../util/got/datasource');
+const got = require('../../util/got');
 const github = require('../github');
 const { parse } = require('../../util/purl');
