+++ lib/datasource/gradle-version/index.js
@@ -1,5 +1,5 @@
 const { coerce } = require('semver');
-const got = require('../../util/got/datasource');
+const got = require('../../util/got');
 
 module.exports = {
   getPkgReleases,
