+++ lib/datasource/maven/index.js
@@ -3,7 +3,7 @@ const fs = require('fs-extra');
 const { XmlDocument } = require('xmldoc');
 const is = require('@sindresorhus/is');
 
-const got = require('../../util/got/datasource');
+const got = require('../../util/got');
 const { compare } = require('../../versioning/maven/compare');
 
 module.exports = {
