+++ lib/datasource/packagist/index.js
@@ -4,7 +4,7 @@ const delay = require('delay');
 const parse = require('github-url-from-git');
 const pAll = require('p-all');
 
-const got = require('../../util/got/datasource');
+const got = require('../../util/got');
 const hostRules = require('../../util/host-rules');
 
 module.exports = {
