+++ lib/datasource/pypi/index.js
@@ -2,7 +2,7 @@ const url = require('url');
 const is = require('@sindresorhus/is');
 
 const { matches } = require('../../versioning/pep440');
-const got = require('../../util/got/datasource');
+const got = require('../../util/got');
 
 module.exports = {
   getPkgReleases,
