+++ lib/datasource/ruby-version/index.js
@@ -1,6 +1,6 @@
 const { parse } = require('node-html-parser');
 
-const got = require('../../util/got/datasource');
+const got = require('../../util/got');
 const { isVersion } = require('../../versioning/ruby');
 
 module.exports = {
