+++ lib/datasource/rubygems/get.js
@@ -1,6 +1,6 @@
 const URL = require('url');
 
-const got = require('../../util/got/datasource');
+const got = require('../../util/got');
 const { maskToken } = require('../../util/mask');
 const hostRules = require('../../util/host-rules');
 const retriable = require('./retriable');
