+++ lib/datasource/terraform/index.js
@@ -1,7 +1,7 @@
 const parse = require('github-url-from-git');
 const is = require('@sindresorhus/is');
 
-const got = require('../../util/got/datasource');
+const got = require('../../util/got');
 
 module.exports = {
   getPkgReleases,
