+++ lib/util/got/cache-get.js
@@ -1,6 +1,9 @@
 const crypto = require('crypto');
 const got = require('got');
 
+// global.repoCache is reset to {} every time a repository is initialized
+// With this caching, it means every GET request is cached during each repository run
+
 module.exports = got.create({
   options: {},
   handler: (options, next) => {
