+++ lib/util/got/renovate-agent.js
@@ -1,5 +1,7 @@
 const got = require('got');
 
+// Sets the user agent to be Renovate
+
 module.exports = got.extend({
   headers:
     process.env.RENOVATE_USER_AGENT ||
