+++ test/datasource/cargo.spec.js
@@ -1,6 +1,6 @@
 const fs = require('fs');
 
-const got = require('../../lib/util/got/datasource');
+const got = require('../../lib/util/got');
 const { getPkgReleases } = require('../../lib/datasource/cargo');
 
 let res1 = fs.readFileSync('test/_fixtures/cargo/libc.json', 'utf8');
@@ -8,7 +8,7 @@ res1 = JSON.parse(res1);
 let res2 = fs.readFileSync('test/_fixtures/cargo/amethyst.json', 'utf8');
 res2 = JSON.parse(res2);
 
-jest.mock('../../lib/util/got/datasource');
+jest.mock('../../lib/util/got');
 
 describe('datasource/cargo', () => {
   describe('getPkgReleases', () => {
