+++ test/datasource/docker.spec.js
@@ -1,9 +1,9 @@
-const got = require('../../lib/util/got/datasource');
+const got = require('../../lib/util/got');
 const docker = require('../../lib/datasource/docker');
 const { getPkgReleases } = require('../../lib/datasource');
 const hostRules = require('../../lib/util/host-rules');
 
-jest.mock('../../lib/util/got/datasource');
+jest.mock('../../lib/util/got');
 jest.mock('../../lib/util/host-rules');
 
 describe('api/docker', () => {
